Separated widgets in other file , directory widgets
